feat(recipe): edit create update recipe ingredient dto to use ingredient id

Closes #66

See merge request scrum-team-4/scrum-backend!43

+++ src/main/java/edu/ntnu/idatt2106/smartmat/dto/recipe/RecipeIngredientDTO.java
@@ -8,6 +8,12 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
 
+/**
+ * Data transfer object for a recipe ingredient.
+ * Displays an ingredient with ingredient data
+ * @author Simen G., Carl G.
+ * @version 1.1 - 26.04.2021
+ */
 @Data
 @Builder
 @NoArgsConstructor
@@ -19,5 +25,5 @@ public class RecipeIngredientDTO {
   private IngredientDTO ingredient;
 
   @NotBlank
-  private int amount;
+  private double amount;
 }
